+++ src/routes.js
@@ -1,12 +1,19 @@
+/* eslint-disable no-unused-vars */
 // Este es el punto de entrada de tu aplicacion
 // eslint-disable-next-line import/no-cycle
 import { home } from './lib/home.js';
 import { login } from './lib/login.js';
 import { post } from './lib/post.js';
 // eslint-disable-next-line import/no-cycle
-import { makingPost } from './app.js';
+// import { makingPost } from './app.js';
 import { signIn } from './lib/signIn.js';
 import { signUp } from './lib/signUp.js';
+
+let firebase = null;
+
+export const loadDependencies = (firebaseFromApp) => {
+  firebase = firebaseFromApp;
+};
 
 export const rootDiv = document.getElementById('root');
 
@@ -40,6 +47,7 @@ const addButtonEvents = () => {
     btn.addEventListener('click', (e) => {
       e.preventDefault();
       const click = e.target.dataset.action;
+      // eslint-disable-next-line no-console
       console.log(click);
       // eslint-disable-next-line no-use-before-define
       eventsController(click);
@@ -47,8 +55,39 @@ const addButtonEvents = () => {
   });
 };
 
+export const makingPost = () => {
+  const titleCard = document.getElementById('title');
+  const subtitleCard = document.getElementById('subtitle');
+  const bodyCard = document.getElementById('body');
+
+  // postButton.addEventListener('click', (e) => {
+  //   e.preventDefault();
+
+  const post = {
+    title: titleCard.value,
+    subtitle: subtitleCard.value,
+    body: bodyCard.value,
+    fecha: Date.now(),
+  };
+
+  if (!titleCard.value.trim() || !subtitleCard.value.trim() || !bodyCard.value.trim()) {
+    // eslint-disable-next-line no-alert
+    alert('Input vacío!');
+    return;
+  }
+
+  savePost(post)
+    .then((docRef) => {
+      console.log('Document written whith ID: ', docRef.id);
+      titleCard.value = '';
+      subtitleCard.value = '';
+      bodyCard.value = '';
+    })
+    .catch((error) => console.log(error));
+};
+
 // Esta es la aplicación que genera el routing
-const eventsController = (e) => {
+export const eventsController = (e) => {
   // eslint-disable-next-line default-case
   switch (e) {
     case 'home':
