+++ src/style.css
@@ -2,23 +2,7 @@ body{
     background: #f9f9f9;
 }
 
-
-a{
-    font-family: "Lufga", sans-serif;
-    text-decoration: none;
-    font-size: 14px;
-    color: black;
-}
-
-a:hover{
-    color: #56F89A;
-}
-
-li{
-    list-style: none;
-}
-
-h1, h2, h3, h6{
+h1{
     font-family: "Lufga", sans-serif;
     padding-top: 20px;
     margin-left: 25px;
@@ -36,6 +20,7 @@ h3{
     font-size: 13px;
     font-weight: 200;
     padding-top: 0px;
+    padding-left: 15px;
 }
 
 h6{
@@ -68,38 +53,9 @@ textarea{
     box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.05);
 }
 
-body > ul:nth-child(2) {
+/* body > ul:nth-child(2) {
     padding-left: 15px;
-}
-
-
-
-/*HAMBURGER MENU*/
-.enabled-menu{
-    display: block;
-  }
-  
-  .disabled-menu{
-    display: none;
-  }
-
- #btn{
-    width: 40px;
-    background-color: #e8e8e8;
-    border: none;
-    color: blue;
-    margin-top: 30px;
-    margin-left: 30px;
-    position: absolute;
-    border-radius: 5px;
-  }  
-
-  .navTittle {
-    position: absolute;
-    margin-top: 80px;
-    display: flex;
-    background: #e9e9;/* revisar posición*/
-  }
+} */
 
   #logIn {
     /*! margin-top: ; */
@@ -219,6 +175,7 @@ margin-top: 32px;
   width: 100%;
   margin-bottom: 0px;
   margin-top: 24px;
+  padding-left: 15px;
    }
 
 #cardSubtitle {
@@ -338,10 +295,10 @@ margin-top: 32px;
     width: 15PX;
     height: 15px;
     position: absolute;
-    float: ;
+    
     top: 700px;
     left: 293px;
-} */
+}
 
 #title {
   margin-left: 38px;
@@ -350,7 +307,7 @@ margin-top: 32px;
   font-family: "Lufga", sans-serif;
   font-size: 14px;
   margin-top: 16px;
-  margin-left: 36px;
+  
   background-color: white;
   border: none;
   border-radius: 0.5em;
@@ -382,4 +339,4 @@ margin-top: 32px;
     border: 1px solid #56F89A;
     font-size: 14px;
     box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.15);
-}
+};
